Combined Util.java and QuickBooksUtil.java
