+++ app/src/test/java/invoice_automation/QuickBooksUtilTest.java
@@ -1,23 +1,28 @@
 package invoice_automation;
 
 import com.intuit.ipp.data.Customer;
+import com.intuit.ipp.data.EmailAddress;
 import com.intuit.ipp.data.Invoice;
 import com.intuit.ipp.data.Item;
 import com.intuit.ipp.data.Line;
 import com.intuit.ipp.data.LineDetailTypeEnum;
+import com.intuit.ipp.data.PhysicalAddress;
 import com.intuit.ipp.data.ReferenceType;
 import com.intuit.ipp.data.SalesItemLineDetail;
+import com.intuit.ipp.data.TelephoneNumber;
+import invoice_automation.model.Address;
 import invoice_automation.model.Conference;
 import invoice_automation.model.InvoiceType;
 import invoice_automation.model.ItemType;
 import invoice_automation.model.PaymentMethod;
+import invoice_automation.model.School;
+import invoice_automation.utils.QuickBooksUtil;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 
@@ -28,22 +33,16 @@ import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
 
 @RunWith(PowerMockRunner.class)
-public class UtilTest {
+public class QuickBooksUtilTest {
     private static final String CUSTOMER_ID = "ID";
     private static final String CUSTOMER_ID1 = "id";
     private static final String SCHOOL_NAME = "Cal";
     private static final String ITEM_ID = "itemid";
-    private static final String ITEM_ID1 = "ITEM ID";
     private static final String ITEM_NAME = "bonus item";
-    private static final String ITEM_NAME1 = "bonus item 2";
     private static final BigDecimal NUM_DELEGATES = BigDecimal.valueOf(24);
-    private static final BigDecimal DEL_FEE_AMT = BigDecimal.valueOf(85);
+    private static final BigDecimal FC_DEL_FEE_AMT = BigDecimal.valueOf(35);
+    private static final BigDecimal BMUN_DEL_FEE_AMT = BigDecimal.valueOf(85);
     private static final BigDecimal CREDIT_CARD_FEE_AMT = BigDecimal.valueOf(34.3);
-    // TODO get rid of this and change references to refer to a value in Consts
-    private static final BigDecimal FC_SCHOOL_FEE_AMT = BigDecimal.valueOf(25);
-    // TODO: Change these to Dates
-    private static final LocalDate INVOICE_DATE = LocalDate.of(2023, 11, 6);
-    private static final LocalDate DUE_DATE = LocalDate.of(2023, 12, 15);
 
     @Mock
     private Invoice invoice;
@@ -62,6 +61,15 @@ public class UtilTest {
     @Mock
     private Item item;
 
+    private final List<String> phoneNumbers = List.of("1234567890", "0987654321");
+    private final Address address = new Address("110 Sproul Hall", "", "Berkeley", "CA", "US", "94720");
+    private final School school = School.builder()
+            .schoolName("Berkeley")
+            .address(address)
+            .email("oski@berkeley.edu")
+            .phoneNumbers(phoneNumbers)
+            .build();
+
     @Test
     public void testCheckInvoiceMatchesCustomer_happyPath_invoiceAndCustomerMatch() {
         // Setup
@@ -70,7 +78,7 @@ public class UtilTest {
         when(customer.getId()).thenReturn(CUSTOMER_ID);
 
         // Do
-        boolean match = Util.checkInvoiceMatchesCustomer(invoice, customer);
+        boolean match = QuickBooksUtil.checkInvoiceMatchesCustomer(invoice, customer);
 
         // Check
         assertTrue(match);
@@ -84,7 +92,7 @@ public class UtilTest {
         when(customer.getId()).thenReturn(CUSTOMER_ID1);
 
         // Do
-        boolean match = Util.checkInvoiceMatchesCustomer(invoice, customer);
+        boolean match = QuickBooksUtil.checkInvoiceMatchesCustomer(invoice, customer);
 
         // Check
         assertFalse(match);
@@ -96,7 +104,7 @@ public class UtilTest {
         when(invoice.getCustomerRef()).thenReturn(null);
 
         // Do
-        boolean match = Util.checkInvoiceMatchesCustomer(invoice, customer);
+        boolean match = QuickBooksUtil.checkInvoiceMatchesCustomer(invoice, customer);
 
         // Check
         assertFalse(match);
@@ -113,16 +121,16 @@ public class UtilTest {
 
         // Do, Check
         when(itemRef.getName()).thenReturn(InvoiceType.BMUN_SCHOOL_FEE.toString());
-        assertEquals(InvoiceType.BMUN_SCHOOL_FEE, Util.getInvoiceTypeFromInvoice(invoice));
+        assertEquals(InvoiceType.BMUN_SCHOOL_FEE, QuickBooksUtil.getInvoiceTypeFromInvoice(invoice));
 
         when(itemRef.getName()).thenReturn(InvoiceType.BMUN_DELEGATE_FEE.toString());
-        assertEquals(InvoiceType.BMUN_DELEGATE_FEE, Util.getInvoiceTypeFromInvoice(invoice));
+        assertEquals(InvoiceType.BMUN_DELEGATE_FEE, QuickBooksUtil.getInvoiceTypeFromInvoice(invoice));
 
         when(itemRef.getName()).thenReturn(InvoiceType.FC_SCHOOL_FEE.toString());
-        assertEquals(InvoiceType.FC_SCHOOL_FEE, Util.getInvoiceTypeFromInvoice(invoice));
+        assertEquals(InvoiceType.FC_SCHOOL_FEE, QuickBooksUtil.getInvoiceTypeFromInvoice(invoice));
 
         when(itemRef.getName()).thenReturn(InvoiceType.FC_DELEGATE_FEE.toString());
-        assertEquals(InvoiceType.FC_DELEGATE_FEE, Util.getInvoiceTypeFromInvoice(invoice));
+        assertEquals(InvoiceType.FC_DELEGATE_FEE, QuickBooksUtil.getInvoiceTypeFromInvoice(invoice));
     }
 
     @Test
@@ -136,7 +144,7 @@ public class UtilTest {
         when(itemRef.getName()).thenReturn("UCBMUN");
 
         // Do, Check
-        assertNull(Util.getInvoiceTypeFromInvoice(invoice));
+        assertNull(QuickBooksUtil.getInvoiceTypeFromInvoice(invoice));
     }
 
     @Test
@@ -147,7 +155,7 @@ public class UtilTest {
         when(line2.getDetailType()).thenReturn(LineDetailTypeEnum.TAX_LINE_DETAIL);
 
         // Do, Check
-        assertNull(Util.getInvoiceTypeFromInvoice(invoice));
+        assertNull(QuickBooksUtil.getInvoiceTypeFromInvoice(invoice));
     }
 
     @Test
@@ -157,7 +165,7 @@ public class UtilTest {
         when(customer.getDisplayName()).thenReturn(SCHOOL_NAME);
 
         // Run
-        ReferenceType customerRef = Util.getCustomerRefFromCustomer(customer);
+        ReferenceType customerRef = QuickBooksUtil.getCustomerRefFromCustomer(customer);
 
         // Verify
         ReferenceType expectedCustomerRef = new ReferenceType();
@@ -173,7 +181,7 @@ public class UtilTest {
         when(item.getName()).thenReturn(ITEM_NAME);
 
         // Run
-        ReferenceType itemRef = Util.getItemRefFromItem(item);
+        ReferenceType itemRef = QuickBooksUtil.getItemRefFromItem(item);
 
         // Verify
         ReferenceType expectedItemRef = new ReferenceType();
@@ -183,91 +191,10 @@ public class UtilTest {
     }
 
     @Test
-    public void testGetRegistrationRoundFromRegistrationDate() {
-        // TODO
-        assert false;
-    }
-
-    @Test
-    public void testCalculateCreditCardProcessingFee() {
-        // TODO
-        assert false;
-    }
-
-    @Test
-    public void testConstructInvoice_happyPath() {
-        // Setup
-        // Construct what would be needed for a mock BMUN del fee invoice
-        ReferenceType customerRef = new ReferenceType();
-        customerRef.setName(SCHOOL_NAME);
-        customerRef.setValue(CUSTOMER_ID);
-        Map<ItemType, BigDecimal> itemQuantityMap = Map.of(
-                ItemType.BMUN_DELEGATE_FEE,
-                NUM_DELEGATES,
-                ItemType.CREDIT_CARD_PROCESSING_FEE,
-                BigDecimal.valueOf(1)
-        );
-        Map<ItemType, BigDecimal> itemRateMap = Map.of(
-                ItemType.BMUN_DELEGATE_FEE,
-                DEL_FEE_AMT,
-                ItemType.CREDIT_CARD_PROCESSING_FEE,
-                CREDIT_CARD_FEE_AMT
-        );
-        ReferenceType itemRef1 = new ReferenceType();
-        itemRef1.setName(ITEM_NAME);
-        itemRef1.setValue(ITEM_ID);
-        ReferenceType itemRef2 = new ReferenceType();
-        itemRef2.setName(ITEM_NAME1);
-        itemRef2.setValue(ITEM_ID1);
-        Map<ItemType, ReferenceType> itemRefMap =
-                Map.of(ItemType.BMUN_DELEGATE_FEE, itemRef1, ItemType.CREDIT_CARD_PROCESSING_FEE, itemRef2);
-
-        // Run
-        // TODO: fix call to pass date
-        /*
-        Invoice invoice = Util.constructInvoice(
-                customerRef,
-                itemQuantityMap,
-                itemRateMap,
-                itemRefMap,
-                PaymentMethod.CARD
-        )
-         */
-        ;
-
-        // Verify
-        // Construct expected Invoice
-        Invoice expectedInvoice = new Invoice();
-        expectedInvoice.setCustomerRef(customerRef);
-        // TODO: set invoice date, due date
-        expectedInvoice.setAllowOnlineCreditCardPayment(true);
-        // Construct the lines
-        Line delFeeLineItem = new Line();
-        delFeeLineItem.setDetailType(LineDetailTypeEnum.SALES_ITEM_LINE_DETAIL);
-        delFeeLineItem.setAmount(DEL_FEE_AMT.multiply(NUM_DELEGATES));
-        SalesItemLineDetail delFeeLineDetail = new SalesItemLineDetail();
-        delFeeLineDetail.setItemRef(itemRef1);
-        delFeeLineDetail.setQty(NUM_DELEGATES);
-        delFeeLineDetail.setUnitPrice(DEL_FEE_AMT);
-        delFeeLineItem.setSalesItemLineDetail(delFeeLineDetail);
-        Line creditCardFeeItem = new Line();
-        creditCardFeeItem.setDetailType(LineDetailTypeEnum.SALES_ITEM_LINE_DETAIL);
-        creditCardFeeItem.setAmount(CREDIT_CARD_FEE_AMT);
-        SalesItemLineDetail creditCardFeeLineDetail = new SalesItemLineDetail();
-        creditCardFeeLineDetail.setItemRef(itemRef2);
-        creditCardFeeLineDetail.setQty(BigDecimal.valueOf(1));
-        creditCardFeeLineDetail.setUnitPrice(CREDIT_CARD_FEE_AMT);
-        creditCardFeeItem.setSalesItemLineDetail(creditCardFeeLineDetail);
-        expectedInvoice.setLine(List.of(delFeeLineItem, creditCardFeeItem));
-
-        assertEquals(expectedInvoice, invoice);
-    }
-
-    @Test
     public void testConstructItemQuantityMap_schoolFeeWithCard() {
         // Run
         Map<ItemType, BigDecimal> itemQuantityMap =
-                Util.constructItemQuantityMap(Conference.BMUN, PaymentMethod.CARD, 0);
+                QuickBooksUtil.constructItemQuantityMap(Conference.BMUN, PaymentMethod.CARD, 0);
 
         // Verify
         Map<ItemType, BigDecimal> expectedMap = Map.of(
@@ -283,7 +210,7 @@ public class UtilTest {
     public void testConstructItemQuantityMap_delegateFeeNoCard() {
         // Run
         Map<ItemType, BigDecimal> itemQuantityMap =
-                Util.constructItemQuantityMap(Conference.BMUN, PaymentMethod.CHECK, NUM_DELEGATES.intValue());
+                QuickBooksUtil.constructItemQuantityMap(Conference.BMUN, PaymentMethod.CHECK, NUM_DELEGATES.intValue());
 
         // Verify
         Map<ItemType, BigDecimal> expectedMap =
@@ -295,12 +222,12 @@ public class UtilTest {
     public void testConstructItemRateMap_delegateFeeWithCard() {
         // Run
         Map<ItemType, BigDecimal> itemRateMap =
-                Util.constructItemRateMap(Conference.BMUN, false, CREDIT_CARD_FEE_AMT);
+                QuickBooksUtil.constructItemRateMap(Conference.BMUN, false, CREDIT_CARD_FEE_AMT);
 
         // Verify
         Map<ItemType, BigDecimal> expectedMap = Map.of(
                 ItemType.BMUN_DELEGATE_FEE,
-                DEL_FEE_AMT,
+                BMUN_DEL_FEE_AMT,
                 ItemType.CREDIT_CARD_PROCESSING_FEE,
                 CREDIT_CARD_FEE_AMT
         );
@@ -311,11 +238,11 @@ public class UtilTest {
     public void testConstructItemRateMap_schoolFeeNoCard() {
         // Run
         Map<ItemType, BigDecimal> itemRateMap =
-                Util.constructItemRateMap(Conference.FC, true, BigDecimal.valueOf(0));
+                QuickBooksUtil.constructItemRateMap(Conference.FC, true, BigDecimal.valueOf(0));
 
         // Verify
         Map<ItemType, BigDecimal> expectedMap =
-                Map.of(ItemType.FC_SCHOOL_FEE, DEL_FEE_AMT);
+                Map.of(ItemType.FC_SCHOOL_FEE, FC_DEL_FEE_AMT);
         assertEquals(expectedMap, itemRateMap);
     }
 
@@ -336,7 +263,7 @@ public class UtilTest {
 
         // Run
         Map<ItemType, ReferenceType> itemRefMap =
-                Util.constructItemRefMap(allRefsMap, true, PaymentMethod.CARD, Conference.BMUN);
+                QuickBooksUtil.constructItemRefMap(allRefsMap, true, PaymentMethod.CARD, Conference.BMUN);
 
         // Verify
         ReferenceType schoolFeeItemRef = allRefsMap.get(ItemType.BMUN_SCHOOL_FEE);
@@ -355,11 +282,37 @@ public class UtilTest {
 
         // Run
         Map<ItemType, ReferenceType> itemRefMap =
-                Util.constructItemRefMap(allRefsMap, false, PaymentMethod.CHECK, Conference.FC);
+                QuickBooksUtil.constructItemRefMap(allRefsMap, false, PaymentMethod.CHECK, Conference.FC);
 
         // Verify
         ReferenceType delFeeItemRef = allRefsMap.get(ItemType.FC_DELEGATE_FEE);
         Map<ItemType, ReferenceType> expectedMap = Map.of(ItemType.FC_DELEGATE_FEE, delFeeItemRef);
         assertEquals(expectedMap, itemRefMap);
+    }
+    @Test
+    public void getCustomerFromSchool_happyPath() {
+        // Run
+        Customer customer = QuickBooksUtil.getCustomerFromSchool(school);
+
+        // Verify
+        Customer expectedCustomer = new Customer();
+        expectedCustomer.setCompanyName("Berkeley");
+        expectedCustomer.setDisplayName("Berkeley");
+        expectedCustomer.setPrimaryEmailAddr(new EmailAddress());
+        expectedCustomer.getPrimaryEmailAddr().setAddress("oski@berkeley.edu");
+        PhysicalAddress schoolAddress = new PhysicalAddress();
+        schoolAddress.setLine1("110 Sproul Hall");
+        schoolAddress.setLine2("");
+        schoolAddress.setCity("Berkeley");
+        schoolAddress.setCountrySubDivisionCode("CA");
+        schoolAddress.setCountry("US");
+        schoolAddress.setPostalCode("94720");
+        expectedCustomer.setBillAddr(schoolAddress);
+        expectedCustomer.setShipAddr(schoolAddress);
+        expectedCustomer.setPrimaryPhone(new TelephoneNumber());
+        expectedCustomer.getPrimaryPhone().setFreeFormNumber(phoneNumbers.get(0));
+        expectedCustomer.setAlternatePhone(new TelephoneNumber());
+        expectedCustomer.getAlternatePhone().setFreeFormNumber(phoneNumbers.get(1));
+        assertEquals(expectedCustomer, customer);
     }
 }
