+++ app/src/test/java/invoice_automation/module/QuickBooksModuleTest.java
@@ -9,13 +9,15 @@ import com.intuit.ipp.data.Item;
 import com.intuit.ipp.data.Line;
 import com.intuit.ipp.data.LineDetailTypeEnum;
 import com.intuit.ipp.data.MemoRef;
+import com.intuit.ipp.data.PhysicalAddress;
 import com.intuit.ipp.data.ReferenceType;
 import com.intuit.ipp.data.SalesItemLineDetail;
+import com.intuit.ipp.data.TelephoneNumber;
 import com.intuit.ipp.exception.FMSException;
 import com.intuit.ipp.services.DataService;
 import com.intuit.ipp.util.Config;
 import invoice_automation.QuickBooksException;
-import invoice_automation.Util;
+import invoice_automation.utils.QuickBooksUtil;
 import invoice_automation.model.Address;
 import invoice_automation.model.Conference;
 import invoice_automation.model.InvoiceType;
@@ -49,7 +51,7 @@ import static org.mockito.Mockito.verify;
 import static org.powermock.api.mockito.PowerMockito.*;
 
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({QuickBooksModule.class, Config.class, Util.class})
+@PrepareForTest({QuickBooksModule.class, Config.class, QuickBooksUtil.class})
 public class QuickBooksModuleTest {
     private static final String ACCESS_TOKEN = "access token";
     private static final String REALM_ID = "realm id";
@@ -92,6 +94,7 @@ public class QuickBooksModuleTest {
             .paymentMethod(PaymentMethod.CARD)
             .conference(Conference.FC)
             .build();
+    private final List<String> phoneNumbers = List.of("1234567890", "0987654321");
 
     @Test
     public void testQuickBooksModuleConstructor_noSandbox_happyPath() {
@@ -321,11 +324,11 @@ public class QuickBooksModuleTest {
         randomInvoice.setId("random fee");
         when(dataService.findAll(any(Invoice.class)))
                 .thenReturn(List.of(schoolFeeInvoice, delFeeInvoice, randomInvoice));
-        mockStatic(Util.class);
-        when(Util.checkInvoiceMatchesCustomer(any(Invoice.class), eq(customer))).thenReturn(true);
-        when(Util.getInvoiceTypeFromInvoice(schoolFeeInvoice)).thenReturn(InvoiceType.BMUN_SCHOOL_FEE);
-        when(Util.getInvoiceTypeFromInvoice(delFeeInvoice)).thenReturn(InvoiceType.BMUN_DELEGATE_FEE);
-        when(Util.getInvoiceTypeFromInvoice(randomInvoice)).thenReturn(null);
+        mockStatic(QuickBooksUtil.class);
+        when(QuickBooksUtil.checkInvoiceMatchesCustomer(any(Invoice.class), eq(customer))).thenReturn(true);
+        when(QuickBooksUtil.getInvoiceTypeFromInvoice(schoolFeeInvoice)).thenReturn(InvoiceType.BMUN_SCHOOL_FEE);
+        when(QuickBooksUtil.getInvoiceTypeFromInvoice(delFeeInvoice)).thenReturn(InvoiceType.BMUN_DELEGATE_FEE);
+        when(QuickBooksUtil.getInvoiceTypeFromInvoice(randomInvoice)).thenReturn(null);
 
         // Do
         Map<InvoiceType, Invoice> invoiceMap = quickBooksModule.queryInvoicesFromRegistration(registration);
@@ -368,8 +371,8 @@ public class QuickBooksModuleTest {
         randomInvoice.setId("random fee");
         when(dataService.findAll(any(Invoice.class)))
                 .thenReturn(List.of(schoolFeeInvoice, delFeeInvoice, randomInvoice));
-        mockStatic(Util.class);
-        when(Util.checkInvoiceMatchesCustomer(any(Invoice.class), eq(customer))).thenReturn(false);
+        mockStatic(QuickBooksUtil.class);
+        when(QuickBooksUtil.checkInvoiceMatchesCustomer(any(Invoice.class), eq(customer))).thenReturn(false);
 
         // Do
         Map<InvoiceType, Invoice> invoiceMap = quickBooksModule.queryInvoicesFromRegistration(registration);
